backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
lxml==5.2.2
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
python-docx==1.1.2
sqlparse==0.5.1
typing_extensions==4.12.2
